+++ tests/testthat/test-conditional_formatting.R
@@ -595,7 +595,6 @@ test_that("extend dataBar tests", {
     params = list(showValue = TRUE, gradient = FALSE)
   )
 
-
   # chained test with rule
   wb$
     add_data(x = -5:5, start_col = 11)$
@@ -610,30 +609,263 @@ test_that("extend dataBar tests", {
   exp <- data.frame(
     sqref = c("A1:A11", "C1:C11", "E1:E11", "G1:G11", "I1:I11", "K1:K11"),
     cf = c(
-      "<cfRule type=\"dataBar\" priority=\"1\"><dataBar showValue=\"1\"><cfvo type=\"min\"/><cfvo type=\"max\"/><color rgb=\"FF638EC6\"/></dataBar><extLst><ext uri=\"{B025F937-C7B1-47D3-B67F-A62EFF666E3E}\" xmlns:x14=\"http://schemas.microsoft.com/office/spreadsheetml/2009/9/main\"><x14:id>{F7189283-14F7-4DE0-9601-54DE9DB40000}</x14:id></ext></extLst></cfRule>",
-      "<cfRule type=\"dataBar\" priority=\"2\"><dataBar showValue=\"0\"><cfvo type=\"min\"/><cfvo type=\"max\"/><color rgb=\"FF638EC6\"/></dataBar><extLst><ext uri=\"{B025F937-C7B1-47D3-B67F-A62EFF666E3E}\" xmlns:x14=\"http://schemas.microsoft.com/office/spreadsheetml/2009/9/main\"><x14:id>{F7189283-14F7-4DE0-9601-54DE9DB40001}</x14:id></ext></extLst></cfRule>",
-      "<cfRule type=\"dataBar\" priority=\"3\"><dataBar showValue=\"0\"><cfvo type=\"min\"/><cfvo type=\"max\"/><color rgb=\"FFA6A6A6\"/></dataBar><extLst><ext uri=\"{B025F937-C7B1-47D3-B67F-A62EFF666E3E}\" xmlns:x14=\"http://schemas.microsoft.com/office/spreadsheetml/2009/9/main\"><x14:id>{F7189283-14F7-4DE0-9601-54DE9DB40002}</x14:id></ext></extLst></cfRule>",
-      "<cfRule type=\"dataBar\" priority=\"4\"><dataBar showValue=\"1\"><cfvo type=\"min\"/><cfvo type=\"max\"/><color rgb=\"FFFF0000\"/></dataBar><extLst><ext uri=\"{B025F937-C7B1-47D3-B67F-A62EFF666E3E}\" xmlns:x14=\"http://schemas.microsoft.com/office/spreadsheetml/2009/9/main\"><x14:id>{F7189283-14F7-4DE0-9601-54DE9DB40003}</x14:id></ext></extLst></cfRule>",
-      "<cfRule type=\"dataBar\" priority=\"5\"><dataBar showValue=\"1\"><cfvo type=\"min\"/><cfvo type=\"max\"/><color rgb=\"FFA6A6A6\"/></dataBar><extLst><ext uri=\"{B025F937-C7B1-47D3-B67F-A62EFF666E3E}\" xmlns:x14=\"http://schemas.microsoft.com/office/spreadsheetml/2009/9/main\"><x14:id>{F7189283-14F7-4DE0-9601-54DE9DB40004}</x14:id></ext></extLst></cfRule>",
-      "<cfRule type=\"dataBar\" priority=\"6\"><dataBar showValue=\"1\"><cfvo type=\"num\" val=\"0\"/><cfvo type=\"num\" val=\"5\"/><color rgb=\"FFA6A6A6\"/></dataBar><extLst><ext uri=\"{B025F937-C7B1-47D3-B67F-A62EFF666E3E}\" xmlns:x14=\"http://schemas.microsoft.com/office/spreadsheetml/2009/9/main\"><x14:id>{F7189283-14F7-4DE0-9601-54DE9DB40005}</x14:id></ext></extLst></cfRule>"
+      '<cfRule type="dataBar" priority="1"><dataBar showValue="1"><cfvo type="min"/><cfvo type="max"/><color rgb="FF638EC6"/></dataBar><extLst><ext uri="{B025F937-C7B1-47D3-B67F-A62EFF666E3E}" xmlns:x14="http://schemas.microsoft.com/office/spreadsheetml/2009/9/main"><x14:id>{F7189283-14F7-4DE0-9601-54DE9DB40000}</x14:id></ext></extLst></cfRule>',
+      '<cfRule type="dataBar" priority="2"><dataBar showValue="0"><cfvo type="min"/><cfvo type="max"/><color rgb="FF638EC6"/></dataBar><extLst><ext uri="{B025F937-C7B1-47D3-B67F-A62EFF666E3E}" xmlns:x14="http://schemas.microsoft.com/office/spreadsheetml/2009/9/main"><x14:id>{F7189283-14F7-4DE0-9601-54DE9DB40001}</x14:id></ext></extLst></cfRule>',
+      '<cfRule type="dataBar" priority="3"><dataBar showValue="0"><cfvo type="min"/><cfvo type="max"/><color rgb="FFA6A6A6"/></dataBar><extLst><ext uri="{B025F937-C7B1-47D3-B67F-A62EFF666E3E}" xmlns:x14="http://schemas.microsoft.com/office/spreadsheetml/2009/9/main"><x14:id>{F7189283-14F7-4DE0-9601-54DE9DB40002}</x14:id></ext></extLst></cfRule>',
+      '<cfRule type="dataBar" priority="4"><dataBar showValue="1"><cfvo type="min"/><cfvo type="max"/><color rgb="FFFF0000"/></dataBar><extLst><ext uri="{B025F937-C7B1-47D3-B67F-A62EFF666E3E}" xmlns:x14="http://schemas.microsoft.com/office/spreadsheetml/2009/9/main"><x14:id>{F7189283-14F7-4DE0-9601-54DE9DB40003}</x14:id></ext></extLst></cfRule>',
+      '<cfRule type="dataBar" priority="5"><dataBar showValue="1"><cfvo type="min"/><cfvo type="max"/><color rgb="FFA6A6A6"/></dataBar><extLst><ext uri="{B025F937-C7B1-47D3-B67F-A62EFF666E3E}" xmlns:x14="http://schemas.microsoft.com/office/spreadsheetml/2009/9/main"><x14:id>{F7189283-14F7-4DE0-9601-54DE9DB40004}</x14:id></ext></extLst></cfRule>',
+      '<cfRule type="dataBar" priority="6"><dataBar showValue="1"><cfvo type="num" val="0"/><cfvo type="num" val="5"/><color rgb="FFA6A6A6"/></dataBar><extLst><ext uri="{B025F937-C7B1-47D3-B67F-A62EFF666E3E}" xmlns:x14="http://schemas.microsoft.com/office/spreadsheetml/2009/9/main"><x14:id>{F7189283-14F7-4DE0-9601-54DE9DB40005}</x14:id></ext></extLst></cfRule>'
     ),
     stringsAsFactors = FALSE
   )
   got <- wb$worksheets[[1]]$conditionalFormatting
-  expect_equal(exp, got)
+  expect_equal(got, exp)
 
-  exp <- read_xml("<ext xmlns:x14=\"http://schemas.microsoft.com/office/spreadsheetml/2009/9/main\" uri=\"{78C0D931-6437-407d-A8EE-F0AAD7539E65}\">
-                  <x14:conditionalFormattings xmlns:xm=\"http://schemas.microsoft.com/office/excel/2006/main\">
-                  <x14:conditionalFormatting><x14:cfRule type=\"dataBar\" id=\"{F7189283-14F7-4DE0-9601-54DE9DB40000}\"><x14:dataBar minLength=\"0\" maxLength=\"100\" border=\"1\" gradient=\"1\" negativeBarBorderColorSameAsPositive=\"0\"><x14:cfvo type=\"autoMin\"/><x14:cfvo type=\"autoMax\"/><x14:borderColor rgb=\"FF638EC6\"/><x14:negativeFillColor rgb=\"FFFF0000\"/><x14:negativeBorderColor rgb=\"FFFF0000\"/><x14:axisColor rgb=\"FF000000\"/></x14:dataBar></x14:cfRule><xm:sqref>A1:A11</xm:sqref></x14:conditionalFormatting>
-                  <x14:conditionalFormatting><x14:cfRule type=\"dataBar\" id=\"{F7189283-14F7-4DE0-9601-54DE9DB40001}\"><x14:dataBar minLength=\"0\" maxLength=\"100\" border=\"1\" gradient=\"0\" negativeBarBorderColorSameAsPositive=\"0\"><x14:cfvo type=\"autoMin\"/><x14:cfvo type=\"autoMax\"/><x14:borderColor rgb=\"FF638EC6\"/><x14:negativeFillColor rgb=\"FFFF0000\"/><x14:negativeBorderColor rgb=\"FFFF0000\"/><x14:axisColor rgb=\"FF000000\"/></x14:dataBar></x14:cfRule><xm:sqref>C1:C11</xm:sqref></x14:conditionalFormatting>
-                  <x14:conditionalFormatting><x14:cfRule type=\"dataBar\" id=\"{F7189283-14F7-4DE0-9601-54DE9DB40002}\"><x14:dataBar minLength=\"0\" maxLength=\"100\" border=\"1\" gradient=\"1\" negativeBarBorderColorSameAsPositive=\"0\"><x14:cfvo type=\"autoMin\"/><x14:cfvo type=\"autoMax\"/><x14:borderColor rgb=\"FFA6A6A6\"/><x14:negativeFillColor rgb=\"FFFF0000\"/><x14:negativeBorderColor rgb=\"FFFF0000\"/><x14:axisColor rgb=\"FF000000\"/></x14:dataBar></x14:cfRule><xm:sqref>E1:E11</xm:sqref></x14:conditionalFormatting>
-                  <x14:conditionalFormatting><x14:cfRule type=\"dataBar\" id=\"{F7189283-14F7-4DE0-9601-54DE9DB40003}\"><x14:dataBar minLength=\"0\" maxLength=\"100\" border=\"1\" gradient=\"0\" negativeBarBorderColorSameAsPositive=\"0\"><x14:cfvo type=\"autoMin\"/><x14:cfvo type=\"autoMax\"/><x14:borderColor rgb=\"FFFF0000\"/><x14:negativeFillColor rgb=\"FFFF0000\"/><x14:negativeBorderColor rgb=\"FFFF0000\"/><x14:axisColor rgb=\"FF000000\"/></x14:dataBar></x14:cfRule><xm:sqref>G1:G11</xm:sqref></x14:conditionalFormatting>
-                  <x14:conditionalFormatting><x14:cfRule type=\"dataBar\" id=\"{F7189283-14F7-4DE0-9601-54DE9DB40004}\"><x14:dataBar minLength=\"0\" maxLength=\"100\" border=\"1\" gradient=\"0\" negativeBarBorderColorSameAsPositive=\"0\"><x14:cfvo type=\"autoMin\"/><x14:cfvo type=\"autoMax\"/><x14:borderColor rgb=\"FFA6A6A6\"/><x14:negativeFillColor rgb=\"FFA6A6A6\"/><x14:negativeBorderColor rgb=\"FFA6A6A6\"/><x14:axisColor rgb=\"FF000000\"/></x14:dataBar></x14:cfRule><xm:sqref>I1:I11</xm:sqref></x14:conditionalFormatting>
-                  <x14:conditionalFormatting><x14:cfRule type=\"dataBar\" id=\"{F7189283-14F7-4DE0-9601-54DE9DB40005}\"><x14:dataBar minLength=\"0\" maxLength=\"100\" border=\"1\" gradient=\"0\" negativeBarBorderColorSameAsPositive=\"0\"><x14:cfvo type=\"num\"><xm:f>0</xm:f></x14:cfvo><x14:cfvo type=\"num\"><xm:f>5</xm:f></x14:cfvo><x14:borderColor rgb=\"FFA6A6A6\"/><x14:negativeFillColor rgb=\"FFA6A6A6\"/><x14:negativeBorderColor rgb=\"FFA6A6A6\"/><x14:axisColor rgb=\"FF000000\"/></x14:dataBar></x14:cfRule><xm:sqref>K1:K11</xm:sqref></x14:conditionalFormatting>
-                  </x14:conditionalFormattings></ext>", pointer = FALSE)
+  exp <- read_xml(
+    '
+    <ext xmlns:x14="http://schemas.microsoft.com/office/spreadsheetml/2009/9/main" uri="{78C0D931-6437-407d-A8EE-F0AAD7539E65}">
+     <x14:conditionalFormattings xmlns:xm="http://schemas.microsoft.com/office/excel/2006/main">
+      <x14:conditionalFormatting>
+       <x14:cfRule type="dataBar" id="{F7189283-14F7-4DE0-9601-54DE9DB40000}">
+        <x14:dataBar minLength="0" maxLength="100" border="1" negativeBarBorderColorSameAsPositive="0">
+         <x14:cfvo type="autoMin" />
+         <x14:cfvo type="autoMax" />
+         <x14:borderColor rgb="FF638EC6" />
+         <x14:negativeFillColor rgb="FFFF0000" />
+         <x14:negativeBorderColor rgb="FFFF0000" />
+         <x14:axisColor rgb="FF000000" />
+        </x14:dataBar>
+       </x14:cfRule>
+       <xm:sqref>A1:A11</xm:sqref>
+      </x14:conditionalFormatting>
+      <x14:conditionalFormatting>
+       <x14:cfRule type="dataBar" id="{F7189283-14F7-4DE0-9601-54DE9DB40001}">
+        <x14:dataBar minLength="0" maxLength="100" showValue="0" border="1" gradient="0" negativeBarBorderColorSameAsPositive="0">
+         <x14:cfvo type="autoMin" />
+         <x14:cfvo type="autoMax" />
+         <x14:borderColor rgb="FF638EC6" />
+         <x14:negativeFillColor rgb="FFFF0000" />
+         <x14:negativeBorderColor rgb="FFFF0000" />
+         <x14:axisColor rgb="FF000000" />
+        </x14:dataBar>
+       </x14:cfRule>
+       <xm:sqref>C1:C11</xm:sqref>
+      </x14:conditionalFormatting>
+      <x14:conditionalFormatting>
+       <x14:cfRule type="dataBar" id="{F7189283-14F7-4DE0-9601-54DE9DB40002}">
+        <x14:dataBar minLength="0" maxLength="100" showValue="0" border="1" negativeBarBorderColorSameAsPositive="0">
+         <x14:cfvo type="autoMin" />
+         <x14:cfvo type="autoMax" />
+         <x14:borderColor rgb="FFA6A6A6" />
+         <x14:negativeFillColor rgb="FFFF0000" />
+         <x14:negativeBorderColor rgb="FFFF0000" />
+         <x14:axisColor rgb="FF000000" />
+        </x14:dataBar>
+       </x14:cfRule>
+       <xm:sqref>E1:E11</xm:sqref>
+      </x14:conditionalFormatting>
+      <x14:conditionalFormatting>
+       <x14:cfRule type="dataBar" id="{F7189283-14F7-4DE0-9601-54DE9DB40003}">
+        <x14:dataBar minLength="0" maxLength="100" border="1" gradient="0" negativeBarBorderColorSameAsPositive="0">
+         <x14:cfvo type="autoMin" />
+         <x14:cfvo type="autoMax" />
+         <x14:borderColor rgb="FFFF0000" />
+         <x14:negativeFillColor rgb="FFFF0000" />
+         <x14:negativeBorderColor rgb="FFFF0000" />
+         <x14:axisColor rgb="FF000000" />
+        </x14:dataBar>
+       </x14:cfRule>
+       <xm:sqref>G1:G11</xm:sqref>
+      </x14:conditionalFormatting>
+      <x14:conditionalFormatting>
+       <x14:cfRule type="dataBar" id="{F7189283-14F7-4DE0-9601-54DE9DB40004}">
+        <x14:dataBar minLength="0" maxLength="100" border="1" gradient="0" negativeBarBorderColorSameAsPositive="0">
+         <x14:cfvo type="autoMin" />
+         <x14:cfvo type="autoMax" />
+         <x14:borderColor rgb="FFA6A6A6" />
+         <x14:negativeFillColor rgb="FFA6A6A6" />
+         <x14:negativeBorderColor rgb="FFA6A6A6" />
+         <x14:axisColor rgb="FF000000" />
+        </x14:dataBar>
+       </x14:cfRule>
+       <xm:sqref>I1:I11</xm:sqref>
+      </x14:conditionalFormatting>
+      <x14:conditionalFormatting>
+       <x14:cfRule type="dataBar" id="{F7189283-14F7-4DE0-9601-54DE9DB40005}">
+        <x14:dataBar minLength="0" maxLength="100" border="1" gradient="0" negativeBarBorderColorSameAsPositive="0">
+         <x14:cfvo type="num">
+          <xm:f>0</xm:f>
+         </x14:cfvo>
+         <x14:cfvo type="num">
+          <xm:f>5</xm:f>
+         </x14:cfvo>
+         <x14:borderColor rgb="FFA6A6A6" />
+         <x14:negativeFillColor rgb="FFA6A6A6" />
+         <x14:negativeBorderColor rgb="FFA6A6A6" />
+         <x14:axisColor rgb="FF000000" />
+        </x14:dataBar>
+       </x14:cfRule>
+       <xm:sqref>K1:K11</xm:sqref>
+      </x14:conditionalFormatting>
+     </x14:conditionalFormattings>
+    </ext>
+    ',
+    pointer = FALSE
+  )
   got <- wb$worksheets[[1]]$extLst
-  expect_equal(exp, got)
+  expect_equal(got, exp)
+})
 
+test_that("test more databar parameters", {
+  vec <- -5:5
+
+  dim1 <- wb_dims(x = vec)
+  dim2 <- wb_dims(x = vec, from_dims = dim1, right = 2)
+  dim3 <- wb_dims(x = vec, from_dims = dim2, right = 2)
+  dim4 <- wb_dims(x = vec, from_dims = dim3, right = 2)
+  dim5 <- wb_dims(x = vec, from_dims = dim4, right = 2)
+  dim6 <- wb_dims(x = vec, from_dims = dim5, right = 2)
+  dim7 <- wb_dims(x = vec, from_dims = dim6, right = 2)
+
+  wb <- wb_workbook()$
+    add_worksheet()$
+    add_data(x = vec, dims = dim1)$
+    add_conditional_formatting(
+      dims = dim1, type = "dataBar", params = list(border = FALSE)
+    )
+
+  wb$add_data(x = vec, dims = dim2)$
+    add_conditional_formatting(
+      dims = dim2, type = "dataBar", style = "#0000FF", params = list(negativeBarColorSameAsPositive = TRUE)
+    )
+
+  wb$add_data(x = vec, dims = dim3)$
+    add_conditional_formatting(
+      dims = dim3, type = "dataBar", params = list(axisPosition = "none")
+    )
+
+  wb$add_data(x = vec, dims = dim4)$
+    add_conditional_formatting(
+      dims = dim4, type = "dataBar", params = list(axisPosition = "middle", axisColor = "FF00FF00")
+    )
+
+  wb$add_data(x = vec, dims = dim5)$
+    add_conditional_formatting(
+      dims = dim5, type = "dataBar", params = list(direction = "rightToLeft", gradient = FALSE)
+    )
+
+  wb$add_data(x = vec, dims = dim6)$
+    add_conditional_formatting(
+      dims = dim6, type = "dataBar", style = c("#0000FF", "#FF0000"),
+      params = list(border = TRUE, negativeBarBorderColorSameAsPositive = TRUE)
+    )
+
+  wb$add_data(x = vec, dims = dim7)$
+    add_conditional_formatting(
+      dims = dim7, type = "dataBar",
+      params = list(minLength = 10, maxLength = 90)
+    )
+
+  exp <- read_xml('
+<ext xmlns:x14="http://schemas.microsoft.com/office/spreadsheetml/2009/9/main" uri="{78C0D931-6437-407d-A8EE-F0AAD7539E65}">
+ <x14:conditionalFormattings xmlns:xm="http://schemas.microsoft.com/office/excel/2006/main">
+  <x14:conditionalFormatting>
+   <x14:cfRule type="dataBar" id="{F7189283-14F7-4DE0-9601-54DE9DB40000}">
+    <x14:dataBar minLength="0" maxLength="100" negativeBarBorderColorSameAsPositive="0">
+     <x14:cfvo type="autoMin" />
+     <x14:cfvo type="autoMax" />
+     <x14:negativeFillColor rgb="FFFF0000" />
+     <x14:axisColor rgb="FF000000" />
+    </x14:dataBar>
+   </x14:cfRule>
+   <xm:sqref>A1:A11</xm:sqref>
+  </x14:conditionalFormatting>
+  <x14:conditionalFormatting>
+   <x14:cfRule type="dataBar" id="{F7189283-14F7-4DE0-9601-54DE9DB40001}">
+    <x14:dataBar minLength="0" maxLength="100" border="1" negativeBarColorSameAsPositive="1" negativeBarBorderColorSameAsPositive="0">
+     <x14:cfvo type="autoMin" />
+     <x14:cfvo type="autoMax" />
+     <x14:borderColor rgb="FF0000FF" />
+     <x14:negativeBorderColor rgb="FFFF0000" />
+     <x14:axisColor rgb="FF000000" />
+    </x14:dataBar>
+   </x14:cfRule>
+   <xm:sqref>C1:C11</xm:sqref>
+  </x14:conditionalFormatting>
+  <x14:conditionalFormatting>
+   <x14:cfRule type="dataBar" id="{F7189283-14F7-4DE0-9601-54DE9DB40002}">
+    <x14:dataBar minLength="0" maxLength="100" border="1" axisPosition="none" negativeBarBorderColorSameAsPositive="0">
+     <x14:cfvo type="autoMin" />
+     <x14:cfvo type="autoMax" />
+     <x14:borderColor rgb="FF638EC6" />
+     <x14:negativeFillColor rgb="FFFF0000" />
+     <x14:negativeBorderColor rgb="FFFF0000" />
+    </x14:dataBar>
+   </x14:cfRule>
+   <xm:sqref>E1:E11</xm:sqref>
+  </x14:conditionalFormatting>
+  <x14:conditionalFormatting>
+   <x14:cfRule type="dataBar" id="{F7189283-14F7-4DE0-9601-54DE9DB40003}">
+    <x14:dataBar minLength="0" maxLength="100" border="1" axisPosition="middle" negativeBarBorderColorSameAsPositive="0">
+     <x14:cfvo type="autoMin" />
+     <x14:cfvo type="autoMax" />
+     <x14:borderColor rgb="FF638EC6" />
+     <x14:negativeFillColor rgb="FFFF0000" />
+     <x14:negativeBorderColor rgb="FFFF0000" />
+     <x14:axisColor rgb="FF00FF00" />
+    </x14:dataBar>
+   </x14:cfRule>
+   <xm:sqref>G1:G11</xm:sqref>
+  </x14:conditionalFormatting>
+  <x14:conditionalFormatting>
+   <x14:cfRule type="dataBar" id="{F7189283-14F7-4DE0-9601-54DE9DB40004}">
+    <x14:dataBar minLength="0" maxLength="100" border="1" gradient="0" direction="rightToLeft" negativeBarBorderColorSameAsPositive="0">
+     <x14:cfvo type="autoMin" />
+     <x14:cfvo type="autoMax" />
+     <x14:borderColor rgb="FF638EC6" />
+     <x14:negativeFillColor rgb="FFFF0000" />
+     <x14:negativeBorderColor rgb="FFFF0000" />
+     <x14:axisColor rgb="FF000000" />
+    </x14:dataBar>
+   </x14:cfRule>
+   <xm:sqref>I1:I11</xm:sqref>
+  </x14:conditionalFormatting>
+  <x14:conditionalFormatting>
+   <x14:cfRule type="dataBar" id="{F7189283-14F7-4DE0-9601-54DE9DB40005}">
+    <x14:dataBar minLength="0" maxLength="100" border="1">
+     <x14:cfvo type="autoMin" />
+     <x14:cfvo type="autoMax" />
+     <x14:borderColor rgb="FFFF0000" />
+     <x14:negativeFillColor rgb="FF0000FF" />
+     <x14:axisColor rgb="FF000000" />
+    </x14:dataBar>
+   </x14:cfRule>
+   <xm:sqref>K1:K11</xm:sqref>
+  </x14:conditionalFormatting>
+  <x14:conditionalFormatting>
+   <x14:cfRule type="dataBar" id="{F7189283-14F7-4DE0-9601-54DE9DB40006}">
+    <x14:dataBar border="1" negativeBarBorderColorSameAsPositive="0">
+     <x14:cfvo type="autoMin" />
+     <x14:cfvo type="autoMax" />
+     <x14:borderColor rgb="FF638EC6" />
+     <x14:negativeFillColor rgb="FFFF0000" />
+     <x14:negativeBorderColor rgb="FFFF0000" />
+     <x14:axisColor rgb="FF000000" />
+    </x14:dataBar>
+   </x14:cfRule>
+   <xm:sqref>M1:M11</xm:sqref>
+  </x14:conditionalFormatting>
+ </x14:conditionalFormattings>
+</ext>
+', pointer = FALSE)
+  got <- wb$worksheets[[1]]$extLst
+  expect_equal(got, exp)
+
+  wb <- wb_workbook()$add_worksheet()
+  expect_error(
+    wb$add_conditional_formatting(type = "dataBar", params = list(icon_set = TRUE)),
+    "Invalid parameters: icon_set"
+  )
 })
 
 test_that("wb_conditional_formatting", {
@@ -651,6 +883,61 @@ test_that("wb_conditional_formatting", {
   exp <- data.frame(
     sqref = "A1:A11",
     cf = "<cfRule type=\"dataBar\" priority=\"1\"><dataBar showValue=\"1\"><cfvo type=\"min\"/><cfvo type=\"max\"/><color rgb=\"FF638EC6\"/></dataBar><extLst><ext uri=\"{B025F937-C7B1-47D3-B67F-A62EFF666E3E}\" xmlns:x14=\"http://schemas.microsoft.com/office/spreadsheetml/2009/9/main\"><x14:id>{F7189283-14F7-4DE0-9601-54DE9DB40000}</x14:id></ext></extLst></cfRule>",
+    stringsAsFactors = FALSE
+  )
+  got <- wb$worksheets[[1]]$conditionalFormatting
+  expect_equal(exp, got)
+})
+
+test_that("wb_conditional_formatting - data bar: `axisPosition` works as it should", {
+  wb <- wb_workbook()
+  wb$add_worksheet("dataBar")
+  example_df <- data.frame(a = sample(c(1, 2, -2, -10, -50, 2, 25, 40, 5, -30)))
+
+  # middle
+  wb$add_data("dataBar", example_df, dims = "A1")
+  wb <- wb_add_conditional_formatting(
+    wb,
+    dims = "A2:A11",
+    type = "dataBar",
+    style = c("#001affff", "#ffccccff"),
+    params = list(axisPosition = "middle")
+  )
+  # none
+  wb$add_data("dataBar", example_df, dims = "B1")
+  wb <- wb_add_conditional_formatting(
+    wb,
+    dims = "B2:B11",
+    type = "dataBar",
+    style = c("#001affff", "#ffccccff"),
+    params = list(axisPosition = "none")
+  )
+  # automatic
+  wb$add_data("dataBar", example_df, dims = "C1")
+  wb <- wb_add_conditional_formatting(
+    wb,
+    dims = "C2:C11",
+    type = "dataBar",
+    style = c("#001affff", "#ffccccff"),
+    params = list(axisPosition = "automatic")
+  )
+  # automatic (default)
+  wb$add_data("dataBar", example_df, dims = "D1")
+  wb <- wb_add_conditional_formatting(
+    wb,
+    dims = "D2:D11",
+    type = "dataBar",
+    style = c("#001affff", "#ffccccff")
+  )
+
+  exp <- data.frame(
+    sqref = c("A2:A11", "B2:B11", "C2:C11", "D2:D11"),
+    cf = c(
+      '<cfRule type="dataBar" priority="1"><dataBar showValue="1"><cfvo type="min"/><cfvo type="max"/><color rgb="FFCCCCFF"/></dataBar><extLst><ext uri="{B025F937-C7B1-47D3-B67F-A62EFF666E3E}" xmlns:x14="http://schemas.microsoft.com/office/spreadsheetml/2009/9/main"><x14:id>{F7189283-14F7-4DE0-9601-54DE9DB40000}</x14:id></ext></extLst></cfRule>',
+      '<cfRule type="dataBar" priority="2"><dataBar showValue="1"><cfvo type="min"/><cfvo type="max"/><color rgb="FFCCCCFF"/></dataBar><extLst><ext uri="{B025F937-C7B1-47D3-B67F-A62EFF666E3E}" xmlns:x14="http://schemas.microsoft.com/office/spreadsheetml/2009/9/main"><x14:id>{F7189283-14F7-4DE0-9601-54DE9DB40001}</x14:id></ext></extLst></cfRule>',
+      '<cfRule type="dataBar" priority="3"><dataBar showValue="1"><cfvo type="min"/><cfvo type="max"/><color rgb="FFCCCCFF"/></dataBar><extLst><ext uri="{B025F937-C7B1-47D3-B67F-A62EFF666E3E}" xmlns:x14="http://schemas.microsoft.com/office/spreadsheetml/2009/9/main"><x14:id>{F7189283-14F7-4DE0-9601-54DE9DB40002}</x14:id></ext></extLst></cfRule>',
+      '<cfRule type="dataBar" priority="4"><dataBar showValue="1"><cfvo type="min"/><cfvo type="max"/><color rgb="FFCCCCFF"/></dataBar><extLst><ext uri="{B025F937-C7B1-47D3-B67F-A62EFF666E3E}" xmlns:x14="http://schemas.microsoft.com/office/spreadsheetml/2009/9/main"><x14:id>{F7189283-14F7-4DE0-9601-54DE9DB40003}</x14:id></ext></extLst></cfRule>'
+    ),
     stringsAsFactors = FALSE
   )
   got <- wb$worksheets[[1]]$conditionalFormatting